requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
Unipath==1.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-bootstrap-form==0.2
django-grappelli==2.6.1
django-localeurl==2.0.1
django-rosetta==0.7.4
django-slim==0.7.4
django-tinymce==1.5.2
gunicorn==18.0
polib==1.0.4
psycopg2==2.5.3
pystache==0.5.4
python-decouple==2.2
pytz==2014.4
requests==2.4.3
six==1.6.1
sorl-thumbnail==11.12
static==1.0.2
wsgiref==0.1.2
